refactor(policies): reconcile policy domain naming
Rename-only cleanup so the bare `Policy` noun names the storage record,
matching the `Cipher`/`Folder` convention. No behavioural change;
registry name stays "Policy" (no data migration).

- PolicyView (record)      -> Policy
- Policy (trait)           -> PolicyDefinition
- ResolvedPolicyView<P>    -> PolicyView<P>
- EnforcedPolicy<P>        -> PolicyDecision<P>
- EnforcedPolicyErased     -> PolicyDecisionErased
