+++ crates/bitwarden-policies/src/lib.rs
@@ -7,13 +7,13 @@ mod uniffi_support;
 
 mod models;
 mod policies;
-pub mod policy;
+mod policy;
 mod policy_client;
+mod policy_definition;
 mod policy_sync_handler;
 mod policy_type;
-mod policy_view;
 
-pub use models::{EnforcedPolicyErased, OrganizationUserPolicyContext};
+pub use models::{OrganizationUserPolicyContext, PolicyDecisionErased};
 // Policy structs will be referenced by other crates once this starts being used
 #[allow(unused)]
 pub(crate) use policies::*;
@@ -23,8 +23,8 @@ pub use policies::{
     OrganizationUserNotificationPolicyData, PasswordGeneratorPolicyData, PasswordGeneratorType,
     ResetPasswordPolicyData, SendOptionsPolicyData, VaultTimeoutAction, VaultTimeoutType,
 };
-pub(crate) use policy::Policy;
+pub use policy::{Policy, PolicyId, PolicyParseError};
 pub use policy_client::{PoliciesClientExt, PolicyClient};
+pub(crate) use policy_definition::PolicyDefinition;
 pub use policy_sync_handler::PolicySyncHandler;
 pub use policy_type::{PolicyDataType, PolicyType};
-pub use policy_view::{PolicyId, PolicyParseError, PolicyView};
