+++ crates/bitwarden-policies/src/policies/block_claimed_domain_account_creation.rs
@@ -1,9 +1,9 @@
-use crate::{Policy, PolicyType, policy_type::PolicyDataType};
+use crate::{PolicyDefinition, PolicyType, policy_type::PolicyDataType};
 
 /// Block Claimed Domain Account Creation policy.
 pub struct BlockClaimedDomainAccountCreationPolicy;
 
-impl Policy for BlockClaimedDomainAccountCreationPolicy {
+impl PolicyDefinition for BlockClaimedDomainAccountCreationPolicy {
     type Data = ();
 
     fn policy_type(&self) -> PolicyType {
