+++ crates/bitwarden-policies/src/policies/disable_personal_vault_export.rs
@@ -1,9 +1,9 @@
-use crate::{Policy, PolicyType, policy_type::PolicyDataType};
+use crate::{PolicyDefinition, PolicyType, policy_type::PolicyDataType};
 
 /// Disable Personal Vault Export policy.
 pub struct DisablePersonalVaultExportPolicy;
 
-impl Policy for DisablePersonalVaultExportPolicy {
+impl PolicyDefinition for DisablePersonalVaultExportPolicy {
     type Data = ();
 
     fn policy_type(&self) -> PolicyType {
