+++ crates/bitwarden-policies/src/policies/disable_send.rs
@@ -1,9 +1,9 @@
-use crate::{Policy, PolicyType, policy_type::PolicyDataType};
+use crate::{PolicyDefinition, PolicyType, policy_type::PolicyDataType};
 
 /// Disable Send policy.
 pub struct DisableSendPolicy;
 
-impl Policy for DisableSendPolicy {
+impl PolicyDefinition for DisableSendPolicy {
     type Data = ();
 
     fn policy_type(&self) -> PolicyType {
