+++ crates/bitwarden-policies/src/policies/single_org.rs
@@ -1,9 +1,9 @@
-use crate::{Policy, PolicyType, policy_type::PolicyDataType};
+use crate::{PolicyDefinition, PolicyType, policy_type::PolicyDataType};
 
 /// Single Organization policy.
 pub struct SingleOrgPolicy;
 
-impl Policy for SingleOrgPolicy {
+impl PolicyDefinition for SingleOrgPolicy {
     type Data = ();
 
     fn policy_type(&self) -> PolicyType {
